Major Update: Navigation to Article and Summary, First Attempt at fixing UI
Removed some modularity by getting rid of the SampleArticles component and
merging it with the InputField component. This fixes the issue with the
component not rerendering. Added buttons to navigate between different
states of the app, one goes to the Article and the other to the Summary.
Fixed the input URL functionality, it accepts valid article urls starting
with 'http://' or 'www.'. Added colors and basic styling to the text and
touchable highlights in order to improve the way the app looks.